fine_tuning/requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.0
dill==0.3.8
dnspython==2.6.1
expecttest==0.2.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.22.2
hypothesis==6.98.10
joblib==1.4.0
jsonpointer==2.1
kiwisolver==1.4.5
matplotlib==3.7.0
mkl-service==2.4.0
multidict==6.0.5
multiprocess==0.70.16
optree==0.10.0
pandas==2.0.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pymongo==4.6.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-etcd==0.4.5
regex==2024.4.16
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sortedcontainers==2.4.0
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.1
torchaudio==2.2.1
torchelastic==0.2.2
torchvision==0.17.1
transformers==4.37.2
triton==2.2.0
types-dataclasses==0.6.6
tzdata==2024.1
xxhash==3.4.1
yarl==1.9.4